requirements.txt
# ShikshaSetu Dependencies - Optimal 2025 Model Stack
# Python 3.11+ Required
# Updated: 2025-11-30
#
# Installation:
#   CPU:      pip install -r requirements.txt
#   GPU:      pip install -r requirements.txt && pip install vllm bitsandbytes
#   Apple M:  pip install -r requirements.txt (use MPS backend)
#
# Optimal Model Stack:
#   - Simplification: meta-llama/Llama-3.2-3B-Instruct
#   - Translation:    ai4bharat/indictrans2-en-indic-1B
#   - Embeddings:     BAAI/bge-m3
#   - Reranker:       BAAI/bge-reranker-v2-m3
#   - OCR:            ucaslcl/GOT-OCR2_0
#   - Validation:     google/gemma-2-2b-it

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# ============================================================================
# TASK QUEUE & CACHING (Celery/Redis for workload management)
# ============================================================================
celery>=5.4.0
redis>=5.2.0
kombu>=5.4.0

# ============================================================================
# AI/ML CORE
# ============================================================================
# PyTorch - install appropriate version for your platform:
#   MPS:     pip install torch torchvision torchaudio
#   CUDA:    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.2.0
transformers>=4.45.0
accelerate>=0.30.0

# Tokenizers
tiktoken>=0.6.0
sentencepiece>=0.2.0

# ============================================================================
# OPTIMAL MODEL STACK (2025)
# ============================================================================

# Embeddings & Retrieval - BGE-M3
sentence-transformers>=3.0.0
FlagEmbedding>=1.2.0

# Translation - IndicTrans2
sacremoses>=0.1.1
indic-nlp-library>=0.92

# OCR - GOT-OCR2 (requires trust_remote_code)
# Model: ucaslcl/GOT-OCR2_0
pymupdf>=1.24.0
Pillow>=10.0.0
pdf2image>=1.17.0

# Audio/TTS
TTS>=0.22.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# ============================================================================
# vLLM PRODUCTION SERVING (CUDA only)
# ============================================================================
# Install separately on GPU systems:
# pip install vllm>=0.6.0
# pip install bitsandbytes>=0.44.0

# HTTP client for vLLM API
httpx>=0.28.0

# ============================================================================
# VECTOR DATABASE
# ============================================================================
pgvector>=0.3.0

# ============================================================================
# SECURITY & AUTH
# ============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.17

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas>=2.2.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.0
langdetect>=1.0.9
einops>=0.7.0

# ============================================================================
# DOCUMENT PROCESSING
# ============================================================================
pypdf>=5.0.0
python-docx>=1.1.0
python-pptx>=1.0.0

# ============================================================================
# HTTP & UTILITIES
# ============================================================================
aiofiles>=24.0.0
python-magic>=0.4.27
python-dotenv>=1.0.0

# ============================================================================
# MONITORING & OBSERVABILITY
# ============================================================================
prometheus-client>=0.21.0
prometheus-fastapi-instrumentator>=7.0.0
sentry-sdk[fastapi]>=2.20.0

# ============================================================================
# UTILITIES
# ============================================================================
pyyaml>=6.0.0
click>=8.1.0
rich>=13.9.0
tqdm>=4.67.0
psutil>=5.9.0

# ============================================================================
# OPTIONAL: GPU-SPECIFIC (install separately)
# ============================================================================
# For CUDA systems, also install:
#   pip install vllm>=0.6.0
#   pip install bitsandbytes>=0.44.0
#   pip install flash-attn>=2.5.0  # requires ninja
#
# For quantization:
#   pip install auto-gptq>=0.7.0
#   pip install autoawq>=0.2.0
